Signal implementation for FreeRTOS environment
This patch is not tested yet.

Signal acts like a simple and lightweight binary-counting semaphore.
Signal is designed in such way that it can be sent during interrupt.
Signal is bound to the concrete task (or thread) and won't stack.
So if task didn't wait for few signals then it will receive only
last one. Perfect stuff for handing interrupt events.