STM32F4: introduce maximum ISR priority
FreeRTOS implies constraints on maximum priority that
user's ISR can have. It can't be greather than
configMAX_SYSCALL_INTERRUPT_PRIORITY

As for CM4 the default priority value that ISR will get is 0,
which means that it will have *highest* possible priority.
That is not what FreeRTOS expects.

Thus default value is introduced, which is 0xff, means that
ISRs will be executed with minimal priority possible.

User is up to redefine this value if needed in his\her
configuration files.