>#258 Electron support with autogeneration and simple UART
This commit includes:
 - few incomplete platform BAT tests as well as unity tests
 - autogeneration of the serial driver configuration
 - UART driver that can do only TX
 - PoC of the new theCore build API, that will be integrated in near
 future

Electron platform now becomes first auto-configured platform. It means
that during the build headers must be generates that contains aliases
and configuration classes based on CMake definitions.

It is possible that this API will be mirrored by the variable-based API
to provide ability to configure theCore via console parameters.